+++ zetaclient/tss/crypto.go
@@ -3,66 +3,21 @@ package tss
 import (
 	"bytes"
 	"encoding/base64"
+	"strings"
 
+	"github.com/btcsuite/btcd/chaincfg/chainhash"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/pkg/errors"
-	"github.com/rs/zerolog"
-	tsscommon "gitlab.com/thorchain/tss/go-tss/common"
 	"gitlab.com/thorchain/tss/go-tss/keysign"
 
 	"github.com/zeta-chain/node/pkg/cosmos"
-	"github.com/zeta-chain/node/zetaclient/logs"
 )
 
 var (
-	testKeySignData    = []byte("hello meta")
 	base64Decode       = base64.StdEncoding.Decode
 	base64DecodeString = base64.StdEncoding.DecodeString
+	base64EncodeString = base64.StdEncoding.EncodeToString
 )
-
-// TestKeySign performs a TSS key-sign test of sample data.
-func TestKeySign(keySigner KeySigner, tssPubKey string, logger zerolog.Logger) error {
-	logger = logger.With().Str(logs.FieldModule, "tss_keysign").Logger()
-
-	hashedData := crypto.Keccak256Hash(testKeySignData)
-
-	logger.Info().
-		Str("keysign.test_data", string(testKeySignData)).
-		Str("keysign.test_data_hash", hashedData.String()).
-		Msg("Performing TSS key-sign test")
-
-	req := keysign.NewRequest(
-		tssPubKey,
-		[]string{base64.StdEncoding.EncodeToString(hashedData.Bytes())},
-		10,
-		nil,
-		Version,
-	)
-
-	res, err := keySigner.KeySign(req)
-	switch {
-	case err != nil:
-		return errors.Wrap(err, "key signing request error")
-	case res.Status != tsscommon.Success:
-		logger.Error().Interface("keysign.fail_blame", res.Blame).Msg("Keysign failed")
-		return errors.Wrapf(err, "key signing is not successful (status %d)", res.Status)
-	case len(res.Signatures) == 0:
-		return errors.New("signatures list is empty")
-	}
-
-	// 32B msg hash, 32B R, 32B S, 1B RC
-	signature := res.Signatures[0]
-
-	logger.Info().Interface("keysign.signature", signature).Msg("Received signature from TSS")
-
-	if _, err = VerifySignature(signature, tssPubKey, hashedData.Bytes()); err != nil {
-		return errors.Wrap(err, "signature verification failed")
-	}
-
-	logger.Info().Msg("TSS key-sign test passed")
-
-	return nil
-}
 
 // VerifySignature checks that keysign.Signature is valid and origins from expected TSS public key.
 // Also returns signature as [65]byte (R, S, V)
@@ -115,4 +70,11 @@ func SignatureToBytes(input keysign.Signature) (sig [65]byte, err error) {
 	}
 
 	return sig, nil
+}
+
+// combineDigests combines the digests
+func combineDigests(digestList []string) []byte {
+	digestConcat := strings.Join(digestList, "")
+	digestBytes := chainhash.DoubleHashH([]byte(digestConcat))
+	return digestBytes.CloneBytes()
 }
